extract proxy stuff in separeta modules
